requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
google-genai
requests
paho-mqtt>=2.0